requirements.txt
captum==0.6.0
comet-ml==3.33.10
matplotlib==3.8.0
numpy==1.25.2
optuna==3.3.0
pandas==2.1.3
Pillow==10.1.0
lightning==2.1.2
PyYAML==6.0.1
scikit_learn==1.3.1
seaborn==0.13.0
torch==2.0.1
torchmetrics==1.1.1
torchvision==0.15.2
